UIEUS-457-restructure-monthpicker
